Add files via upload
Ok - changes to maybe help? Countries and Cities plots to horizontal bar charts, so you can see the label names...changes to YearCalc to show only the years 1990 on and to Final_Report.rmd to clean up some language and change values for US Temp changes to reflect new years ... I still can't get the output from YearCalcDiff to show up. It prints the correct values to the screen, but the output isn't showing in knitted html file. Any ideas?